Cleaned up repo and peformance test
