+++ cli/cliQuestions.js
@@ -2,41 +2,12 @@ const inquirer = require('inquirer');
 const chalk = require('chalk');
 
 const {
-  cliFileMethods, folderDir, fileList, htmlList, startScriptArr,
+  cliFileMethods, folderDir, fileList, htmlList,
 } = require('./cliFileMethods');
 
 inquirer.registerPrompt('autocomplete', require('inquirer-autocomplete-prompt'));
 
 const cliQuestions = [
-  {
-    type: 'list', // creates a selectable list of answers that is selectable with the arrow keys
-    message: 'What functionality would you like to add to your project? (select one)',
-    name: 'choiceInstall',
-    choices: [
-      {
-        name: 'Server-side rendering only',
-      },
-      {
-        name: 'Service worker caching for offline functionality',
-      },
-      {
-        name: 'Server-side rendering with service worker caching for offline functionality',
-      },
-    ],
-  },
-  // {
-  //   type: 'autocomplete',
-  //   name: 'static',
-  //   suggestOnly: true,
-  //   message: chalk.red('Type the start script for your developement server (make sure it is not 5001)'),
-  //   source: cliFileMethods.startScripts,
-  //   validate(answer) {
-  //     if (!folderDir.includes(answer)) {
-  //       return 'Invalid entry point';
-  //     }
-  //     return true;
-  //   },
-  // },
   {
     type: 'autocomplete',
     name: 'static',
@@ -62,28 +33,12 @@ const cliQuestions = [
     },
   },
   {
-    type: 'list',
-    message: chalk.green('Are you using Redux?'),
-    name: 'hasRedux',
-    choices: [
-      {
-        name: 'Yes',
-      },
-      {
-        name: 'No',
-      },
-    ],
-  },
-  {
-    type: 'autocomplete',
-    name: 'store',
-    suggestOnly: true,
-    message: chalk.green('Type the path to your store creator:'),
-    source: cliFileMethods.searchFiles,
-    when: (answers) => (answers.hasRedux === 'Yes') ? true : false,
+    type: 'input',
+    name: 'port',
+    message: chalk.green('Type in the port number of your client side server:'),
     validate(answer) {
-      if (!fileList.includes(answer)) {
-        return 'Invalid entry point';
+      if (answer.length === 0 || typeof JSON.parse(answer) !== 'number') {
+        return 'Enter a valid number';
       }
       return true;
     },
@@ -114,55 +69,6 @@ const cliQuestions = [
       return true;
     },
   },
-  // {
-  //   type: 'input', // type designates what type of prompt the user sees, creates a prompt that takes text as an answer
-  //   name: 'projectName', // name is the property the answer to this prompt is saved as
-  //   message: chalk.red('Input your project name'), // message is what the user sees on screen
-  //   validate(answer) {
-  //     // if a user typed nothing or only spaces, make the user type again
-  //     if (answer.length === 0 || answer.trim().length === 0) {
-  //       return 'You must type your project name';
-  //     }
-  //     return true;
-  //   },
-  // },
-  // {
-  //   type: 'list',
-  //   message: chalk.red('Would you like to run html comparison?'),
-  //   name: 'htmlTest',
-  //   choices: [
-  //     {
-  //       name: 'YES',
-  //     },
-  //     {
-  //       name: 'NO',
-  //     },
-  //   ],
-  // },
-  // {
-  //   type: 'checkbox',  //creates a list of items that can be toggled on or off
-  //   message: 'What features would you like to add to your project? (select multiple)',
-  //   name: 'features',
-  //   choices: [
-  //     new inquirer.Separator('Toggle features on or off (stretch stuff)'),
-  //     {
-  //       name: 'Add service worker children',
-  //     },
-  //     {
-  //       name: 'Add indexDb caching',
-  //     },
-  //     {
-  //       name: 'Code splitting with React Router',
-  //     },
-  //   ],
-  //   validate(answer) {
-  //     // checking to see if the user selected a drop down menu
-  //     if (answer.length === 0) {
-  //       return 'You must make a selection';
-  //     }
-  //     return true;
-  //   },
-  // },
 ];
 
 module.exports = cliQuestions;
