+++ lib/performanceTest/createPrimeServer.js
@@ -6,8 +6,6 @@ const app = express();
 
 const port = 5050;
 
-app.use(express.static('./primessr/img'));
-
 app.get('/', (req, res) => {
   res.sendFile(path.join(__dirname, './primeCompare.html'));
 });
