--- lib/performanceTest/perfTiming.js
@@ -1,14 +1,0 @@
-const extractDataFromPerformanceTiming = (timing, ...dataNames) => {
-  const navigationStart = timing.navigationStart;
-
-  const extractedData = {};
-  dataNames.forEach((name) => {
-    extractedData[name] = timing[name] - navigationStart;
-  });
-
-  return extractedData;
-};
-
-module.exports = {
-  extractDataFromPerformanceTiming,
-};
