--- lib/performanceTest/testPage.js
@@ -1,36 +1,0 @@
-const { extractDataFromPerformanceTiming } = require('./perfTiming');
-
-async function testPage(page) {
-  await page.goto('https://www.cnn.com/');
-
-  const performanceTiming = JSON.parse(
-    await page.evaluate(() => JSON.stringify(window.performance.timing))
-  );
-
-  return extractDataFromPerformanceTiming(
-    performanceTiming,
-    'navigationStart',
-    'unloadEventStart',
-    'unloadEventEnd',
-    'redirectStart',
-    'redirectEnd',
-    'fetchStart',
-    'domainLookupStart',
-    'domainLookupEnd',
-    'connectStart',
-    'connectEnd',
-    'secureConnectionStart',
-    'requestStart',
-    'responseStart',
-    'responseEnd',
-    'domLoading',
-    'domInteractive',
-    'domContentLoadedEventStart',
-    'domContentLoadedEventEnd',
-    'domComplete',
-    'loadEventStart',
-    'loadEventEnd',
-  );
-}
-
-module.exports = testPage;
